Fix video cells in `Discover`
