+++ app/src/main/res/layout/activity_player.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:backgroundTint="@color/blackOther"
+    android:backgroundTint="@color/black"
     tools:ignore="SmallSp">
 
     <com.trx.consumer.common.CommonView
